refactor(sun): the shared service, out of the product
solar.py was the photovoltaic product and a service three unrelated products
read: the pv chain, the wind screening, and the canopy, which lights a stand
with the hourly sun of its own location. The service is now terra/sun/, and
solar.py drops from 1,165 lines to 733.

nasa_power.py is the one reader of the POWER record: endpoint, transport,
request shape, decode. `solar._request` becomes `nasa_power.request` -- it is
the single transport for both the radiation and the meteorology paths, so the
two cannot drift apart in how they handle a failure against the same endpoint,
and it now crosses a package boundary to reach the wind screening. A private
name imported across packages is the one-reader rule written as a lie. Named
for the source, because electrical power is the subject of the modules that
come next and the collision would be unreadable.

record.py is what a fetched record says over time: totals, trend, climatology,
clearness, beam share. position.py is where the sun was and how much beam
energy came from there.

TWO DEFERRED IMPORTS THAT MUST STAY DEFERRED. pvlib inside
position.prepare_hourly and scipy inside record.linear_trend. The canopy calls
prepare_hourly on every run and has no use for a photovoltaic library; hoisting
either would load it on runs that never touch an array, and nothing would fail
to say so. terra/sun/__init__.py is deliberately empty for the same reason: a
convenience re-export makes importing one submodule execute the others.

N_HORIZON_AZIMUTHS travels to position.py without the section header it sat
under, which is horizon prose and stays with the horizon code. It gets the
reason it is on this side instead: a solar azimuth indexes straight into the
horizon array, so the two sector counts have to be one number.

THE ALIASES ARE sun_power, sun_position, sun_record. `record` and `position`
are already local variable names in the test modules that now import them, and
a module shadowed by a local is the failure mode this migration keeps meeting.

The repoint was driven by attribute reads in the parsed source, not by text
search: two of the sites the search found are prose -- canopy_field.py:750 and
energy.py:1748 name these functions in docstrings -- and importing a module to
satisfy a docstring is how an unused import gets written.

DEVIATION FROM THE PLAN. tests/test_solar.py is repointed but not yet split
into a sun-side file. Splitting it here would leave the new file importing
solar anyway, for the two names in one test that do not move until the
photovoltaic commit. It splits there instead.

Verification: pytest 506 passed, 8 skipped. gofmt, go vet and go test -race
clean. A ping through the entry point answers. The guard checks all 45 rewired
attribute reads in terra/actions.py, and the deferred-import test confirms
resolving an action still imports neither pvlib nor scipy.
