+++ sidecar/tests/test_infer.py
@@ -167,16 +167,16 @@ def test_request_positive_admits_zero_only_where_zero_is_a_value():
 
 def test_the_power_cache_key_is_not_finer_than_the_grid_it_keys_on():
     """
-    The key used solar.grid_key, which rounds to 0.01 degrees, about 1 km. Two
+    The key used sun_power.grid_key, which rounds to 0.01 degrees, about 1 km. Two
     AOIs inside one POWER cell then missed each other and each paid the roughly
     23 s hourly fetch, so the reuse the cache states it guarantees did not hold.
     """
-    import solar
+    from terra.sun import nasa_power as sun_power
     import wind
 
     a = (-53.5048, -25.7434)
     b = (-53.5362, -25.5)
-    assert solar.grid_key(*a) != solar.grid_key(*b)
+    assert sun_power.grid_key(*a) != sun_power.grid_key(*b)
     assert wind.grid_key(*a) == wind.grid_key(*b)
     assert actions.power_cell_key(*a) == actions.power_cell_key(*b)
